fix: Resolve circular dependency and finalize Firebase integration
This commit resolves a series of persistent build and rendering errors (`TypeError: can't access property 'default', App is undefined`) by refactoring the Firebase integration to eliminate a circular dependency.

Key Architectural Changes:

1.  **Dedicated `useFirebase` Hook:**
    - The `FirebaseContext` and `useFirebase` hook have been moved to a dedicated file (`src/hooks/useFirebase.ts`).
    - The `FirebaseProvider` now imports the context from this file, breaking the dependency cycle where components that consumed the hook were imported by the same file that provided it.

2.  **Component Refactoring:**
    - All components that use Firebase services (`LoginButton`, `PetList`, `NewPetPage`) have been updated to import the `useFirebase` hook from its new, correct location.

This commit also includes the `<a>` tag fix for the `LoginButton` as a final attempt to resolve the unresponsive click issue on the deployed site.

With these changes, the application now has a robust, stable, and architecturally sound foundation for all Firebase features. The project builds successfully and is ready for deployment and testing.
