+++ src/app/feed/page.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect } from "react";
 import Image from "next/image";
 import Link from "next/link";
-import { useFirebase } from "@/components/FirebaseProvider";
+import { useFirebase } from "@/hooks/useFirebase";
 import { collection, query, onSnapshot, orderBy, DocumentData, Timestamp } from "firebase/firestore";
 
 interface Post {
